Added some unit  tested and corrected bug detected by unit test. All unit test are successful now.

+++ SharpBoostVoronoi/SharpBoostVoronoi/Outputs/Cell.cs
@@ -11,12 +11,12 @@ namespace SharpBoostVoronoi.Output
         /// <summary>
         /// The voronoi cell identifier
         /// </summary>
-        public long Index { get; set; }
+        public int Index { get; set; }
 
         /// <summary>
         /// The index of the source feature
         /// </summary>
-        public long Site { get; set; }
+        public int Site { get; set; }
 
         /// <summary>
         /// True if the cell is made from a point
@@ -31,13 +31,13 @@ namespace SharpBoostVoronoi.Output
         /// <summary>
         /// Indexes of the segment that makes the cell
         /// </summary>
-        public List<long> Segments { get; set; }
+        public List<int> Segments { get; set; }
 
         /// <summary>
         /// Constructor
         /// </summary>
         /// <param name="t">A tuple returned by the CLR wrapper.</param>
-        public Cell(Tuple <long, long, bool, bool, List<long>> t)
+        public Cell(Tuple <int, int, bool, bool, List<int>> t)
         {
             Index = t.Item1;
             Site = t.Item2;
